re-structure send email functionalities
