+++ src/utilities/email/sendemail/index.ts
@@ -1,26 +1,10 @@
 import dotenv from "dotenv";
-import nodemailer from "nodemailer";
 import mO from "@src/messages/constants/others";
 import { constantValuesForEmail } from "@src/utilities/email/sendemail/content";
-import { ObjectEmailAndPortType, ObjectEmailBody } from "@src/types";
+import { createTransporter } from "@src/utilities/email/sendemail/mailtransporter";
+import { ObjectEmailBody } from "@src/types";
 
 dotenv.config();
-
-function createTransporter(emailHostPort: ObjectEmailAndPortType) {
-  const { emailHost, emailPort } = { ...emailHostPort };
-
-  const transporter = nodemailer.createTransport({
-    host: emailHost,
-    port: emailPort,
-    auth: {
-      user: process.env.APP_USER,
-      pass: process.env.APP_PASSWORD,
-    },
-  });
-  return function () {
-    return { transporter, emailHost };
-  };
-}
 
 export const sendMail = async function (_body: ObjectEmailBody) {
   const e = constantValuesForEmail();
